# pkg/server/static/openapi.yaml
openapi: 3.1.0
info:
  title: DevRadar API
  description: >-
    DevRadar tracks how a container image's vulnerabilities and licenses change
    over time by rescanning its SBOM. Submit an SBOM pinned to an image digest;
    DevRadar rescans it on a schedule and records every change as an event.


    All endpoints are authenticated with an API token (`Authorization: Bearer
    dr_...`) minted in the UI. Responses are JSON: success bodies are objects;
    list endpoints add a `next_cursor` when more pages exist; errors are
    `{"error": "..."}`. This document is the source of truth for the API
    contract — the human-readable version is at `/api`.
  version: 0.0.0
servers:
  - url: https://devradar.thingz.io
    description: DevRadar SaaS
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /v1/images:
    get:
      tags: [Images]
      summary: List tracked images
      description: >-
        Returns the tenant's images grouped by repository (one row per image),
        risk-ranked. `min_severity` trims each per-severity breakdown; the
        overall `total` is unaffected.
      parameters:
        - { $ref: "#/components/parameters/MinSeverity" }
        - name: q
          in: query
          description: Filter by repository name substring
          schema: { type: string }
        - name: label
          in: query
          description: Filter to images carrying this grouping label
          schema: { type: string }
        - { $ref: "#/components/parameters/Sort" }
        - { $ref: "#/components/parameters/Dir" }
        - { $ref: "#/components/parameters/Cursor" }
        - { $ref: "#/components/parameters/Limit" }
      responses:
        "200":
          description: A page of images
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      min_severity: { type: string }
                      images:
                        type: array
                        items: { $ref: "#/components/schemas/RepoImage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /v1/sboms/{id}:
    get:
      tags: [SBOMs]
      summary: Get SBOM metadata
      description: Metadata and severity breakdown for one SBOM. `min_severity` trims the breakdown.
      parameters:
        - { $ref: "#/components/parameters/SBOMId" }
        - { $ref: "#/components/parameters/MinSeverity" }
      responses:
        "200":
          description: SBOM metadata
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SBOMDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [SBOMs]
      summary: Archive an SBOM
      description: >-
        Stops tracking an SBOM: it drops from the scan set and image list;
        findings and history are retained. Idempotent.
      parameters:
        - { $ref: "#/components/parameters/SBOMId" }
      responses:
        "204": { description: Archived (or already archived) }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /v1/sboms/{id}/licenses:
    get:
      tags: [Licenses]
      summary: Package licenses for an SBOM
      description: >-
        Per-package license inventory for one SBOM, each classified into an
        obligation category and evaluated against the tenant's compliance
        policy. Violations are listed first.
      parameters:
        - { $ref: "#/components/parameters/SBOMId" }
      responses:
        "200":
          description: Packages with licenses and policy verdicts
          content:
            application/json:
              schema:
                type: object
                properties:
                  packages:
                    type: array
                    items: { $ref: "#/components/schemas/PackageLicense" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/licenses:
    get:
      tags: [Licenses]
      summary: Fleet license rollup
      description: >-
        The tenant's fleet-wide license landscape: distribution by license
        family and obligation category, plus unlicensed and policy-violation
        counts across all active SBOMs.
      responses:
        "200":
          description: Fleet license statistics
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FleetLicenseStats" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        An API token minted in the UI (Tokens & settings). Send as
        `Authorization: Bearer dr_...`.
  parameters:
    SBOMId:
      name: id
      in: path
      required: true
      description: The SBOM id (per-tenant content hash) returned at submission.
      schema: { type: string }
    MinSeverity:
      name: min_severity
      in: query
      description: >-
        Minimum severity floor: `critical`|`high`|`medium`|`low`|`negligible`
        (default: the tenant's setting, usually `medium`). `unknown` is always
        included. Resolved independently per endpoint.
      schema:
        type: string
        enum: [critical, high, medium, low, negligible]
    Sort:
      name: sort
      in: query
      description: Sort key (endpoint-specific; omit for the default).
      schema: { type: string }
    Dir:
      name: dir
      in: query
      description: Sort direction.
      schema: { type: string, enum: [asc, desc] }
    Cursor:
      name: cursor
      in: query
      description: Opaque keyset cursor from a prior response's `next_cursor`.
      schema: { type: string }
    Limit:
      name: limit
      in: query
      description: Max rows (default 100, max 1000).
      schema: { type: integer, minimum: 1, maximum: 1000 }
  responses:
...
    Unauthorized:
      description: Missing or invalid API token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found or not owned by the tenant
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string, description: Human-readable error message }
      required: [error]
    Page:
      type: object
      description: List envelope. `next_cursor` is present only when more pages exist.
      properties:
        next_cursor:
          type: string
          description: Pass back as `?cursor=` for the next page.
...
    SeverityCounts:
      type: object
      description: >-
        Per-severity breakdown. Buckets below the requested threshold are
        omitted; `unknown` is always kept. `total` is the overall count;
        `relevant` sums the visible buckets.
      properties:
        critical: { type: integer }
        high: { type: integer }
        medium: { type: integer }
        low: { type: integer }
        negligible: { type: integer }
        unknown: { type: integer }
        total: { type: integer }
        relevant: { type: integer }
    RepoImage:
      type: object
      properties:
        repository: { type: string }
        sbom_count: { type: integer }
        digest_count: { type: integer }
        versions: { type: array, items: { type: string } }
        latest_at: { type: string, format: date-time }
        counts: { $ref: "#/components/schemas/SeverityCounts" }
        fixable: { type: integer }
        failures: { type: integer }
...
    SBOMDetail:
      type: object
      properties:
        sbom_id: { type: string }
        image_ref: { type: string }
        digest: { type: string }
        format: { type: string }
        spec_version: { type: string }
        tool: { type: string }
        tool_version: { type: string }
        status: { type: string, enum: [active, archived] }
        submitted_at: { type: string, format: date-time }
        generated_at: { type: string, format: date-time }
        counts: { $ref: "#/components/schemas/SeverityCounts" }
...
    PackageLicense:
      type: object
      properties:
        package: { type: string }
        version: { type: string }
        purl: { type: string }
        licenses:
          type: array
          items: { type: string }
          description: Raw SPDX IDs / names / expressions (multi-license preserved).
        category:
          type: string
          enum: [permissive, weak-copyleft, strong-copyleft, proprietary, unknown]
        violation: { type: boolean, description: Fails the tenant's compliance policy }
        reason: { type: string, description: Which denied license triggered the violation }
    LicenseCount:
      type: object
      properties:
        key: { type: string }
        count: { type: integer }
    FleetLicenseStats:
      type: object
      properties:
        families:
          type: array
          items: { $ref: "#/components/schemas/LicenseCount" }
        categories:
          type: array
          items: { $ref: "#/components/schemas/LicenseCount" }
        packages: { type: integer, description: Distinct packages with a license }
        unlicensed: { type: integer, description: Distinct packages with no resolvable license }
        violations: { type: integer, description: Distinct packages failing the policy }
...